I integrated 2 FieldTypes
